requirements.txt
# 브라우저 자동화
playwright>=1.40.0

# 주식 데이터
yfinance>=0.2.36

# 데이터 처리
pandas>=2.1.0
numpy>=1.26.0

# 웹 스크래핑
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# 시각화
plotly>=5.18.0
kaleido>=0.2.1

# 터미널 출력
rich>=13.7.0

# 환경변수
python-dotenv>=1.0.0

# AI 분석 (선택적)
openai>=1.12.0
